Safari-609.1.11.1/LayoutTests/imported/w3c/web-platform-tests/css/css-grid/alignment/grid-alignment-implies-size-change-010-expected.txt

CONSOLE MESSAGE: line 190: No valid data-* attributes found in selector list : .after
XX X
X XXX
X
XX XXX

PASS .before 1 
